chore: move files and create storybook
