+++ packages/app/src/systems/SignUp/components/PinWalletCard/PinWalletCard.tsx
