+++ packages/app/src/systems/SignUp/components/PinWalletCard/index.tsx
@@ -1,0 +1,1 @@
+export * from './PinWalletCard';
