+++ packages/app/src/systems/SignUp/components/PinWalletText/index.tsx
@@ -1,2 +1,1 @@
-export * from './PinWalletCard';
 export * from './PinWalletText';
