+++ packages/app/src/systems/SignUp/components/index.tsx
@@ -4,3 +4,4 @@ export * from './MnemonicRead';
 export * from './MnemonicWrite';
 export * from './SignUpFailed';
 export * from './WalletCreated';
+export * from './PinWalletCard';
